feat(api-v2): Return standoff markup separately from text values (#1307)
This PR enables text markup to be returned as standoff/RDF, separately from text values. Standoff is returned in pages of a maximum size. It also queries standoff in pages internally when returning markup as XML, to reduce the size of SPARQL query results. Existing repositories must be updated; see `upgrade/1307-standoff/README.md` for instructions.