Make integration-tests pickup its own conf from application.conf
Remove -Dconfig.resource=fuseki.conf param from it sbt invocation and replace by a dedicated application.conf in it class pass.

The it application.conf will be merged with the default one.

Remove import of standard application.conf in test.conf as this lead to a cyclic dependency during startup.