refactor: Replace Cardinality .isStricterThan(Cardinality):Boolean with isIncludedIn method
A cardinality includes another cardinality if and only if the other cardinality is included in its entire range of possible values.

The previous method "isStricterThan" assumes a commutative property in the sense that `a isStricterThan b == false` then `b isStricterThan a == true` must hold. However, this is not the case for cardinalities `ZeroOrOne` and `AtLeastOne`. Both cardinalities are "stricter" than the other as their range of values only has a very small intersection. In the future with the introduction of other cardinalities e.g. ExactlyTwo an intersection might not even occur.