feat: Allow setting new Cardinalities if they are more restrictive than the respective Cardinalities of a possibly existing super class (#2397)
All checks are done in the context of a class/property combination on which the new cardinality should be set.

Adds an ontology check for subclasses and super-classes so that the new cardinality is compatible, i.e. :

Ontologycheck:
The new cardinality must be included in the the respective cardinalities of the super-classes
The respective subclasses's cardinalities must be included in the new cardinality
Persistencecheck: A simple sanity check for persistent data which prevents incompatible changes: Either the property is not in use or the new cardinality on the class is included in the old cardinality of that class.
Given the property on the current class is not used the ontology check (1.+2.) is sufficient. Otherwise data might exist that possibly does not comply with the new cardinality. In this PR this is not a deeper check, in a later PR I will add a check which would allow to set a new cardinality in the case the the property on that class is in use and the new cardinality is not included in the old cardinality. In this case it could still be possible that the data complies with the new cardinality. This case is not possible in this PR but should later be enabled.

Co-authored-by: Marcin Procyk <marcin.procyk@dasch.swiss>
Co-authored-by: irinaschubert <irina.schubert@dasch.swiss>