fix: Include folder name in mapping.csv in order to prevent collisions (#91)
Should a file exist in a subfolder with the same name as another file we
have to be able to distinguish these files in the mapping.csv. Include
the folder path inside the import/<project_shortcode> as part of the
original in the mapping.csv resolves this ambiguity.

The new content of the `mapping-<project_shortcode>.csv` file would look
like this:

```
original,derivative
100x100.png,75JdC5Idk6o-QjUEZWoAT3r.jpx
subfolder/100x100.png,1Ty6tYGtiMw-fxm9fbsvCV7.jpx
subfolder/subfolder2/100x100.png,3sILHLNzSfw-rbWPMoGmwEC.jpx
...
```

for an import folder like:

```
import/<project_shortcode>/
         ├── 100x100.png
         │ 
         └── subfolder/
                  ├── 100x100.png
                  │
                  └── subfolder2/
                           └── 100x100.png
```