fix: Improve performance for Gravsearch queries
This is an attempt at a minimally invasive fix for some of the performance issue we face with Gravsearch.

Gravsearch queries contain `FILTER NOT EXISTS` blocks which are used to only retrieve resources which were not marked as deleted.
For some reason this slows down the query substantially.
Replacing this block with a `MINUS` makes simple queries faster.

In general, MINUS is often more efficient than FILTER NOT EXISTS for simple cases because it directly computes the difference between two result sets. It can benefit from query optimization and indexing.

FILTER NOT EXISTS, especially when used with complex subqueries or patterns, may be slower, as it needs to check for the non-existence of a certain pattern within the result set.