fix: Send queries to Fuseki as form data not as URL parameters (#3768)
<!-- Important! Please follow the guidelines for naming Pull Requests:
https://docs.dasch.swiss/latest/developers/contribution/ -->

### Description


This pull request updates how SPARQL queries are sent to the triplestore
service in `TriplestoreServiceLive`. The main change is switching back
from passing query parameters in the URL to sending them in the request
body as form data.
This restores previous behaviour and should resolve the `URI too long`
error we were seeing, in case of very long query strings.