test(dsp-api): Reuse test layers across all e2e and it tests (#3800)
This should speed up the tests as the testcontainers do not have to be
spun up for each testsuite, the prepare takes care that the right data
is loaded and the caches are refreshed for each *Spec.

For more info see
https://zio.dev/reference/test/sharing-layers-between-multiple-files

Whilst the E2E spec were running nicely I had to fix the
ProjectImportServiceSpec and the SipiIT so that it does not come to port
conflicts.

Also I have added logging when starting up the testcontainers and also
commented out the log consumers for the testcontainers in order to
reduce the noise. Some cleanup includes removing the now unused
DspIngestClientMock and inlining the corresponging trait with only a
single implementation now.

On my machine -excluding compilation time - `test-e2e` runs in ~ 3
minutes and `test-it` runs in ~ 3.5 minutes.