fix: Footnote contents get double encoded with SPARQL-safe encoding, but decoded only once (DEV-4880, DEV-4796, DEV-5274) (#3737)
### Description

Resolved issue with double-encoding of input values.

Previously, input was SPARQL-encoded at deserialisation, because the
Twirl-based query construction did not do any input sanitisation.
When moving to the RDF4J query builder, values automatically get
SPARQL-encoded. But because the not all queries are migrated, input
encoding remained in place. This lead to double-encoding in some places.


The fix includes:

- [x] Added a test case in `ResourcesRouteV2E2ESpec` to verify that
apostrophes in footnote content are correctly encoded as `&apos;`
- [x] Added a similar test in `ValuesRouteV2E2ESpec` to ensure
consistent behavior when creating text values with standoff
- [x] Implemented a new `ValueRepo.createValue` method that properly
handles text encoding
- [x] Created a new `CreateValueQueryBuilder` to generate SPARQL queries
with proper encoding
- [x] Improved error handling by replacing direct response string output
with a helper method
- [x] Add test for value comments with apostrophe
- [ ] Add tests for updating values
- [x] Implement `addValueVersion` in the same way as `createValue`
- [x] Ensure all the encodings marked with `// XXX` are unproblematic

---------

Co-authored-by: Raitis Veinbahs <raitis.veinbahs@dasch.swiss>