chore: Remove docs generators for ingest and api (#3807)
Since we do not incorporate the open api yaml for ingest or api in the
documenation anymore and only link to the online version we can safely
remove the code generating the yamls.

This does not affect serving the open api yaml from the services.

<!-- Important! Please follow the guidelines for naming Pull Requests:
https://docs.dasch.swiss/latest/developers/contribution/ -->

### Description

<!-- Please add a short description of the changes -->

<!-- * **What kind of change does this PR introduce?** (Bug fix,
feature, docs update, ...) -->

<!-- * **What is the current behavior?** (You can also link to an open
issue here) -->

<!-- * **What is the new behavior (if this is a feature change)?** -->

<!-- * **Does this PR introduce a breaking change?**
(What changes might users need to make in their application due to this
PR?) -->

<!-- * **Other information**: -->