fix: Gravsearch limiting to project wrongly excludes knora-api entities (#3797)
<!-- Important! Please follow the guidelines for naming Pull Requests:
https://docs.dasch.swiss/latest/developers/contribution/ -->

### Description

Previously, when limiting a Gravsearch (or Gravsearch count) query to a
project, it would only consider resource classes defined in the
ontologies of the project.
This wrongly excluded any classes of a type defined in `knora-base`,
such as `LinkObj` or `Region` etc.

The fix adds the `knora-base` ontology to the ontologies considered for
search when `limitToProject` is set.

---------

Co-authored-by: Christian Kleinbölting <seakayone@users.noreply.github.com>