feat: Replace representationClassIri with an enum and count only non deleted resources (DEV-5581) (#3850)
Resource classes can be subclass of a single *Represention class.
Currently we know of a fixed representations such as StillImage or Text
for example. Replacing the `representationClassIri` with an
[enum](https://github.com/dasch-swiss/dsp-api/pull/3850/files#diff-f3e1dfdb7253efdec505c33418174564bdff59c271ad6ea8b1eb201952ba4a83)
will make it easier for the Frontend to deal with these cases.

Filter the counted resource by removing all "knora-base:isDeleted true".